# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Swagger Doc For Picspop app",
      "version": "1.0.0",
      "title": "Swagger Doc For Picspop app",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "paths": {
...
        "/booking": {
            "post": {
                "tags": [
                    "User"
                ],
                "security": [
                    {
                      "Bearer": []
                    }
                  ],
                "summary": "Booking",
                "description": "Booking",
                "operationId": "booking",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/booking"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/booking"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "security": [
                    {
                      "Bearer": []
                    }
                  ],
                "summary": "profile ",
                "description": "profile",
                "operationId": "profile",
                "responses": {}
            }  
          },
...
    },
...
    "servers": [
      {
        "url": "http://192.168.1.196:7807/api"
      },
      {
        "url": "https://mypicspop.com/api"
      }
    ],
    "components": {
      "securitySchemes": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
          }
    },
      "schemas": {
...
    "booking": {
        "type": "object",
        "required": [],
        "properties": {
            "serviceProviderId": {
                "type": "string"
            },
            "categoryId": {
              "type": "number"
          },
            "locationOwnerId": {
                "type": "string"
            },
            "serviceProviderPrice": {
                "type": "string"
            },
            "locationPrice": {
                "type": "string"
            },
            "cardId": {
                "type": "number"
            },
            "cvc": {
                "type": "string"
            },
            "location": {
              "type": "string"
          },
          "lat": {
              "type": "string"
          },
          "lng": {
              "type": "string"
          },
          "startBookingDate": {
            "type": "string"
        },
        "endBookingDate": {
            "type": "string"
        }
        
        
        }
    },
...
      }
    }
  }